TemporaryDigdagServer needs to run shutdown handler for jacoco agent to flush data

+++ digdag-tests/src/test/java/utils/TemporaryDigdagServer.java
@@ -60,6 +60,7 @@ import static utils.TestUtils.main;
 import static com.google.common.base.Strings.isNullOrEmpty;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Arrays.asList;
+import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
@@ -433,16 +434,30 @@ public class TemporaryDigdagServer
 
     private static void kill(Process p)
     {
-        sendUnixSignal(p, "KILL");
-        p.destroyForcibly();
+        // Send TERM first and wait for a while so that shutdown handler runs
+        // and Jacoco agent can write coverage data.
+        boolean exited = false;
+        if (terminate(p)) {
+            try {
+                exited = p.waitFor(5, SECONDS);
+            }
+            catch (InterruptedException ex) {
+                Thread.currentThread().interrupt();
+                // kill anyways
+            }
+        }
+        if (!exited) {
+            sendUnixSignal(p, "KILL");
+            p.destroyForcibly();
+        }
     }
 
-    private static void terminate(Process p)
+    private static boolean terminate(Process p)
     {
-        sendUnixSignal(p, "TERM");
+        return sendUnixSignal(p, "TERM");
     }
 
-    private static void sendUnixSignal(Process p, String signalName)
+    private static boolean sendUnixSignal(Process p, String signalName)
     {
         if (isUnixProcess(p)) {
             int pid = pid(p);
@@ -455,6 +470,10 @@ public class TemporaryDigdagServer
                     log.warn("command failed: {}", asList(cmd), e);
                 }
             }
+            return true;
+        }
+        else {
+            return false;
         }
     }
